# docs/encuestas-okrean-openapi.yaml
...
openapi: 3.1.0
info:
  title: Encuestas Okrean — API de ingesta (v1, v3 y v4)
  version: '4.1.0'
  summary: Recepción de encuestas de campo desde la app móvil Encuestas Okrean.
  description: |
    Contrato entre la app móvil **Encuestas Okrean** y el backend de flotillas.

    **Se aceptan TRES versiones del cuestionario:** `versionCuestionario: 1` (el
    restaurado, con sus dos ramas por `estado`), `versionCuestionario: 3` (el
    vigente) y `versionCuestionario: 4` (v3 con **catálogo propio en P8**
    —sale `paola_barrera`; entran `paco_nino` y `goyo_castaneda`— más las
    opciones `otro` / `no_sabe_no_contesta` y sus preguntas abiertas; P9 sí
    extiende el catálogo de v3). El servidor despacha por ese entero **antes** de validar
    y cada versión tiene su propio schema; cualquier otro entero recibe
    `422 UNSUPPORTED_VERSION`.

    Módulo **separado** de la ingesta de GeoCampo (`/api/qa-externa/*`): tablas
    propias, API keys propias y CLI de alta propio. Lo único que comparten es el
    portal de revisión y la cuenta del revisor.

    **Reglas duras del contrato:**

    - La respuesta de éxito es SIEMPRE `201` (alta) o `200` (reenvío idempotente),
      las dos con cuerpo `{"idRemoto":"…"}`. **Nunca `204`**: la app necesita el
      `idRemoto` para cerrar el registro local.
    - La idempotencia la da `idLocal` (UNIQUE en base de datos) más un hash
      canónico SHA-256 del contenido sustantivo. El mismo `idLocal` con el mismo
...

...
  contact:
...
  license:
...
servers:
...
security:
  - apiKeyDispositivo: []
...
paths:
  /api/v1/encuestas:
    post:
      tags: [ingesta]
      # El nombre conserva el sufijo V3 del spec 3.0.0 aunque hoy también ingesta
      # v1: renombrarlo rompería cualquier cliente ya generado de ese spec.
      operationId: ingestarEncuestaV3
      summary: Envía una encuesta.
      description: |
        Alta idempotente de una encuesta. El servidor mira `versionCuestionario`,
        valida el payload contra el cuestionario **de esa versión** (v1 o v3),
        calcula el hash canónico de esa misma versión e inserta una sola fila.

        Reintentar tras un timeout es seguro: si la fila ya existía con el mismo
        contenido, la respuesta es `200` con el mismo `idRemoto`.

        El prefijo `/api` **no es opcional**: en el despliegue público Caddy manda
        todo el tráfico a Next, y Next solo reescribe hacia la API las rutas que
        empiezan por `/api`. Sin el prefijo, la petición acaba en una redirección
        al login del portal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              # Las tres versiones del cuestionario, seleccionadas por el `const`
              # de `versionCuestionario` (1, 3 o 4) que lleva cada rama. NO se usa
              # `discriminator`: el keyword exige una propiedad de tipo string y
              # aquí el discriminante es un entero.
              oneOf:
                - $ref: '#/components/schemas/EncuestaV1'
                - $ref: '#/components/schemas/EncuestaV3'
                - $ref: '#/components/schemas/EncuestaV4'
            examples:
              encuestaV3:
                summary: Encuesta completada (v3)
                value:
                  idLocal: '7f9c2a10-4d3b-4a7e-9f21-0c8b5d6e1a44'
                  folioLocal: 'LX-001248'
                  encuestador: 'María López'
                  versionCuestionario: 3
                  estado: completada
                  fechaHoraInicio: '2026-07-30T16:12:04.000Z'
                  fechaHoraFinalizacion: '2026-07-30T16:18:41.000Z'
                  duracionSegundos: 397
                  respuestas:
                    sexo: mujer
                    rangoEdad: '31_45'
                    empresariosConocidos: ['Javier González', 'Rosa Mendoza']
                    politicosConocidos: ['Lalo Ximénez', 'Irineo Molina', 'Fernando Huerta']
                    conoceLalo: 'si'
                    rolLalo: politico_lider_social
                    opinionLalo: buena
                    preferenciaElectoral: lalo_ximenez
                    preferenciaPartido: morena
                    aprobacionPorGobernante:
                      - { gobernante: sheinbaum, calificacion: muy_buena }
                      - { gobernante: jara, calificacion: buena }
                      - { gobernante: huerta, calificacion: buena }
                  ubicacion:
                    disponible: true
                    latitud: 19.432608
                    longitud: -99.133209
                    precisionMetros: 12.4
                    fechaHoraCaptura: '2026-07-30T16:18:39.000Z'
                    permiso: concedido
                    servicioActivo: true
                    esValida: true
                  dispositivo:
                    plataforma: android
                    modelo: 'Moto G54'
                    versionSistema: '14'
                  versionAplicacion: '2.0.0'
              encuestaV4:
                summary: Encuesta completada (v4 con "otro")
                value:
                  idLocal: 'a1b2c3d4-e5f6-47a8-9b0c-1d2e3f4a5b6c'
                  folioLocal: 'LX-001250'
                  encuestador: 'Roberto García'
                  versionCuestionario: 4
                  estado: completada
                  fechaHoraInicio: '2026-07-31T09:30:00.000Z'
                  fechaHoraFinalizacion: '2026-07-31T09:35:20.000Z'
                  duracionSegundos: 320
                  respuestas:
                    sexo: hombre
                    rangoEdad: '18_30'
                    empresariosConocidos: ['Juan Pérez']
                    politicosConocidos: ['Otro político local']
                    conoceLalo: 'si'
                    rolLalo: 'politico_lider_social'
                    opinionLalo: 'regular'
                    preferenciaElectoral: otro
                    preferenciaElectoralOtro: 'Candidato independiente'
                    preferenciaPartido: otro
                    preferenciaPartidoOtro: 'Movimiento local'
                    aprobacionPorGobernante:
                      - { gobernante: sheinbaum, calificacion: regular }
                      - { gobernante: jara, calificacion: mala }
                      - { gobernante: huerta, calificacion: buena }
                  ubicacion:
                    disponible: true
                    latitud: 19.5432
                    longitud: -99.1456
                    precisionMetros: 25.0
                    fechaHoraCaptura: '2026-07-31T09:35:15.000Z'
                    permiso: concedido
                    servicioActivo: true
                    esValida: true
                  dispositivo:
                    plataforma: android
                    modelo: 'Samsung A12'
                    versionSistema: '11'
                  versionAplicacion: '2.1.0'
              encuestaV1:
                summary: Encuesta completada, elegible (v1)
                value:
                  idLocal: '3f2504e0-4f89-41d3-9a0c-0305e82c3301'
                  folioLocal: 'LX-1042'
                  encuestador: 'María López'
                  versionCuestionario: 1
                  estado: completada
                  elegibilidad: elegible
                  fechaHoraInicio: '2026-07-30T10:00:00.000Z'
                  fechaHoraFinalizacion: '2026-07-30T10:06:40.000Z'
                  duracionSegundos: 400
                  respuestas:
                    credencialVigente: 'si'
                    rangoEdad: '30_44'
                    genero: mujer
                    partidoPreferido: morena
                    conocimientoPorPersona:
                      - { persona: lalo_ximenez, nivel: bien }
                      - { persona: laura_estrada, nivel: algo }
                      - { persona: paco_nino, nivel: poco }
                      - { persona: gabriela_delgado, nivel: no_conoce }
                      - { persona: irineo_molina, nivel: no_conoce }
                      - { persona: goyo_castaneda, nivel: no_conoce }
                      - { persona: ernesto_montero, nivel: no_conoce }
                    mediosConocimiento:
                      tipo: respondida
                      medios: [redes_sociales, labor_social]
                    mayorPersonalidad: lalo_ximenez
                    candidatoPreferido: laura_estrada
                  ubicacion:
                    disponible: true
                    latitud: 19.432608
                    longitud: -99.133209
                    precisionMetros: 12.5
                    fechaHoraCaptura: '2026-07-30T10:05:00.000Z'
                    permiso: concedido
                    servicioActivo: true
                    esValida: true
                  dispositivo:
                    plataforma: android
                    modelo: 'Moto G54'
                    versionSistema: '14'
                  versionAplicacion: '1.0.3'
      responses:
        '201':
          description: Encuesta creada. Es la respuesta normal del primer envío.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestaAceptada'
              example:
                idRemoto: 'c1a7e0b2-9d44-4f0a-b7c3-2e5f8a91d033'
        '200':
          description: |
            Reenvío idempotente: el `idLocal` ya existía con el **mismo** contenido
            sustantivo. Devuelve el mismo `idRemoto` de la primera vez y no crea
            otra fila. La app debe tratarlo exactamente igual que el `201`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestaAceptada'
              example:
                idRemoto: 'c1a7e0b2-9d44-4f0a-b7c3-2e5f8a91d033'
        '400':
          description: |
            Cuerpo que no es JSON válido (`BAD_JSON`) o que no es un objeto.
            No reintentable sin corregir el payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                error: 'Cuerpo JSON malformado'
                code: BAD_JSON
                requestId: '0f0c1a7e-…'
        '401':
          description: |
            Falta el header `Authorization`, el formato no es `Bearer <key>`, la
            key no existe o el dispositivo fue revocado (`activo = false`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                error: 'API key inválida o revocada'
                code: UNAUTHORIZED
        '403':
          description: |
            Sin permisos. La ingesta por API key **no** emite este código: recibirlo
            significa que la petición no llegó a este router — típicamente una URL
            base mal configurada que cayó en un montaje protegido por JWT. Corregir
            la URL, no reintentar.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
        '409':
          description: |
            El `idLocal` ya está registrado **con contenido sustantivo distinto**.
            La fila original queda intacta. No sirve reintentar con la misma clave:
            es un `idLocal` reutilizado en el teléfono.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                error: 'idLocal ya registrado con contenido distinto'
                code: CONFLICT
                requestId: '0f0c1a7e-…'
        '413':
          description: |
            El cuerpo supera los 256 KB. El límite se aplica sobre el stream, así
            que también corta las peticiones `chunked` sin `Content-Length`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                error: 'Cuerpo de la petición demasiado grande'
                code: PAYLOAD_TOO_LARGE
                requestId: '0f0c1a7e-…'
        '422':
          description: |
            El JSON es válido pero el contenido no pasa el cuestionario de su
            versión (`VALIDATION_ERROR`), o `versionCuestionario` es un entero
            fuera de {1, 3, 4} (`UNSUPPORTED_VERSION`). **Terminal**: reintentar el
            mismo payload volverá a fallar.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorValidacion'
                  - $ref: '#/components/schemas/ErrorRespuesta'
              examples:
                validacionV3:
                  summary: Fallo del schema v3
                  value:
                    error: 'Datos inválidos'
                    code: VALIDATION_ERROR
                    issues:
                      - field: respuestas.aprobacionPorGobernante
                        message: 'aprobacionPorGobernante debe traer los 3 gobernantes del catálogo v3'
                      - field: ubicacion.esValida
                        message: 'esValida debe ser true exactamente cuando precisionMetros <= 50'
                    requestId: '0f0c1a7e-…'
                validacionV1:
                  summary: Fallo del schema v1
                  value:
                    error: 'Datos inválidos'
                    code: VALIDATION_ERROR
                    issues:
                      - field: respuestas.conocimientoPorPersona
                        message: 'conocimientoPorPersona debe traer las 7 personas del catálogo v1'
                      - field: respuestas.mediosConocimiento.tipo
                        message: 'mediosConocimiento debe venir respondida si conoce al menos a una persona'
                    requestId: '0f0c1a7e-…'
                versionNoSoportada:
                  summary: Cuestionario de una versión que este servidor no habla (p. ej. v2, v5)
                  value:
                    error: 'Versión de cuestionario no soportada'
                    code: UNSUPPORTED_VERSION
                    requestId: '0f0c1a7e-…'
        '429':
          description: |
            Se agotó una de las dos cuotas: la del dispositivo o la de la IP. La
            cabecera `Retry-After` trae los segundos que faltan para la siguiente
            ventana. Reintentable con backoff.
          headers:
            Retry-After:
              description: Segundos hasta que la ventana se reabre.
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                error: 'Demasiadas encuestas desde este dispositivo. Intenta más tarde.'
                code: RATE_LIMITED
        '500':
          description: |
            Error interno. **No** hubo escritura parcial: el alta es un solo INSERT
            atómico. Reintentable con el mismo `idLocal`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                error: 'Error interno del servidor'
                code: INTERNAL_ERROR
                requestId: '0f0c1a7e-…'

    get:
      tags: [diagnóstico]
      operationId: encuestasMetodoNoPermitido
      summary: No permitido — responde 405 con credencial válida.
      description: |
        Existe a propósito, como red de seguridad para una app que se equivoque de
        método: sin él la petición saldría de este router y la atraparían los
        montajes comodín con JWT, que responden **401** — y el móvil lee ese 401
        como "API key inválida" y dispara una reconfiguración innecesaria. Un 405
        dice lo que de verdad pasa: la key sirve, el método no.

        Para probar conectividad, usa `/api/v1/encuestas/ping`.
      responses:
        '405':
          description: Método no permitido (la API key sí es válida).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
              example:
                error: 'Method Not Allowed'
                code: METHOD_NOT_ALLOWED
        '401':
          description: Sin credencial válida (el guard corre antes que el 405).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorRespuesta'
...
components:
  securitySchemes:
    apiKeyDispositivo:
      type: http
      scheme: bearer
      description: |
        API key del dispositivo, emitida por el CLI de alta y mostrada **una sola
        vez** (el servidor guarda solo su hash SHA-256). Va en el header
        `Authorization: Bearer <api key>`. No es un JWT: no expira y no tiene
        claims; se revoca dando de baja el dispositivo.
  schemas:
...
    SexoV3:
      type: string
      description: Sexo declarado.
      enum: [hombre, mujer]
    RangoEdadV3:
      type: string
      description: Rango de edad declarado.
      enum: ['18_30', '31_45', '46_mas']
    RolLaloV3:
      type: string
      description: Rol de Lalo Ximénez declarado.
      enum: [politico_lider_social, empresario, funcionario_publico, no_sabe_no_contesta]
    OpinionLaloV3:
      type: string
      description: Opinión sobre Lalo Ximénez declarada.
      enum: [muy_buena, buena, regular, mala, muy_mala, no_lo_conozco]
...
    RespuestasV3:
      type: object
      additionalProperties: false
      required:
        - sexo
        - rangoEdad
        - empresariosConocidos
        - politicosConocidos
        - conoceLalo
        - rolLalo
        - opinionLalo
        - preferenciaElectoral
        - preferenciaPartido
        - aprobacionPorGobernante
      properties:
        sexo:
          $ref: '#/components/schemas/SexoV3'
        rangoEdad:
          $ref: '#/components/schemas/RangoEdadV3'
        empresariosConocidos:
          type: array
          minItems: 0
          maxItems: 3
          description: |
            Nombres de empresarios conocidos. Texto libre, 1–80 caracteres cada uno.
            Se recorta el espacio sobrante (trim).
          items:
            type: string
            minLength: 1
            maxLength: 80
        politicosConocidos:
          type: array
          minItems: 1
          maxItems: 3
          description: |
            Nombres de políticos conocidos. Texto libre, 1–80 caracteres cada uno.
            Debe traer al menos uno.
          items:
            type: string
            minLength: 1
            maxLength: 80
        conoceLalo:
          type: string
          # Entrecomillados: en YAML 1.1 un `no` a secas se lee como booleano false.
          enum: ['si', 'no']
          description: ¿Conoce a Lalo Ximénez?
        rolLalo:
          $ref: '#/components/schemas/RolLaloV3'
        opinionLalo:
          $ref: '#/components/schemas/OpinionLaloV3'
        preferenciaElectoral:
          $ref: '#/components/schemas/PreferenciaElectoralV3'
        preferenciaPartido:
          $ref: '#/components/schemas/PartidoV3'
        aprobacionPorGobernante:
          type: array
          minItems: 3
          maxItems: 3
          description: |
            Exactamente 3 filas, una por gobernante (sheinbaum, jara, huerta),
            sin repetir. El orden en el payload será normalizado al orden canónico
            antes de hashear.
          items:
            $ref: '#/components/schemas/AprobacionPorGobernanteV3'
...
    EncuestaV3:
      title: Encuesta v3 (cuestionario vigente)
      description: |
        Payload de una encuesta v3. Todos los campos de respuesta son obligatorios
        y no hay ramas: `estado` solo puede ser `completada`.
      allOf:
        - $ref: '#/components/schemas/EncuestaBaseV3'
        - type: object
          required: [versionCuestionario, estado, respuestas]
          properties:
            versionCuestionario:
              type: integer
              const: 3
              description: |
                Versión del cuestionario. Además del `3` se acepta el `1` (ver
                `EncuestaV1`) y el `4` (ver `EncuestaV4`); cualquier otro entero
                responde 422 `UNSUPPORTED_VERSION` (y un valor que no sea entero
                positivo, 422 `VALIDATION_ERROR`).
            estado:
              type: string
              const: completada
              description: Siempre `completada` en v3.
            respuestas:
              $ref: '#/components/schemas/RespuestasV3'
    EncuestaV4:
      title: Encuesta v4 (v3 extendida con opciones "otro")
      description: |
        Payload de una encuesta v4. Idéntico a v3 salvo por los nuevos catálogos de
        preferencias (que incluyen `"otro"` y `"no_sabe_no_contesta"`) y dos campos
        opcionales condicionales para especificar en texto libre. Todos los demás
        campos de respuesta son obligatorios y no hay ramas: `estado` solo puede
        ser `completada`.
      allOf:
        - $ref: '#/components/schemas/EncuestaBaseV3'
        - type: object
          required: [versionCuestionario, estado, respuestas]
          properties:
            versionCuestionario:
              type: integer
              const: 4
              description: |
                Versión del cuestionario. Además del `4` se acepta el `1` (ver
                `EncuestaV1`) y el `3` (ver `EncuestaV3`); cualquier otro entero
                responde 422 `UNSUPPORTED_VERSION`.
            estado:
              type: string
              const: completada
              description: Siempre `completada` en v4.
            respuestas:
              $ref: '#/components/schemas/RespuestasV4'
    EncuestaCompletadaV1:
      title: Encuesta v1 completada (elegible)
      description: |
        El encuestado tiene credencial vigente y respondió P2–P8. Todas las
        respuestas son obligatorias en esta rama.
      allOf:
        - $ref: '#/components/schemas/EncuestaBaseV3'
        - type: object
          required: [versionCuestionario, estado, elegibilidad, respuestas]
          properties:
            versionCuestionario:
              type: integer
              const: 1
            estado:
              type: string
              const: completada
            elegibilidad:
              type: string
              const: elegible
            respuestas:
              $ref: '#/components/schemas/RespuestasCompletadaV1'
    EncuestaNoElegibleV1:
      title: Encuesta v1 no elegible
      description: |
        El encuestado no tiene credencial vigente y la app corta el cuestionario en
        P1.
      allOf:
        - $ref: '#/components/schemas/EncuestaBaseV3'
        - type: object
          required: [versionCuestionario, estado, elegibilidad, respuestas]
          properties:
            versionCuestionario:
              type: integer
              const: 1
            estado:
              type: string
              const: noElegible
            elegibilidad:
              type: string
              const: noElegible
            respuestas:
              $ref: '#/components/schemas/RespuestasNoElegibleV1'
    EncuestaV1:
      title: Encuesta v1 (cuestionario restaurado)
      description: |
        Payload de una encuesta v1 (`versionCuestionario: 1`). Unión discriminada
        por `estado`: la rama `completada` exige P2–P8, la rama `noElegible` las
        prohíbe.

        Además del `1`, el servidor acepta el `3` (ver `EncuestaV3`); cualquier
        otro entero responde 422 `UNSUPPORTED_VERSION` (y un valor que no sea
        entero positivo, 422 `VALIDATION_ERROR`).
      oneOf:
        - $ref: '#/components/schemas/EncuestaCompletadaV1'
        - $ref: '#/components/schemas/EncuestaNoElegibleV1'
      discriminator:
        propertyName: estado
        mapping:
          completada: '#/components/schemas/EncuestaCompletadaV1'
          noElegible: '#/components/schemas/EncuestaNoElegibleV1'

    # ── Respuestas ──────────────────────────────────────────────────────────
    IngestaAceptada:
      type: object
      additionalProperties: false
      required: [idRemoto]
      description: |
        Único cuerpo de éxito, idéntico en 201 y en 200. La app solo necesita
        `idRemoto` para cerrar el registro local.
      properties:
        idRemoto:
          type: string
          format: uuid
          description: Identificador asignado por el servidor. Estable para un mismo `idLocal`.
...
    ErrorRespuesta:
      type: object
      required: [error, code]
      description: Forma común de error de la API.
      properties:
        error:
          type: string
          description: Mensaje legible, seguro para mostrar o registrar.
        code:
          type: string
          description: Código estable para ramificar en el cliente.
          examples:
            - BAD_JSON
            - UNAUTHORIZED
            - FORBIDDEN
            - CONFLICT
            - PAYLOAD_TOO_LARGE
            - VALIDATION_ERROR
            - UNSUPPORTED_VERSION
            - METHOD_NOT_ALLOWED
            - RATE_LIMITED
            - INTERNAL_ERROR
        details:
          description: Detalle opcional del error (presente en algunos 400).
        requestId:
          type: string
          description: |
            Correlaciona la respuesta con la línea del log del servidor. Inclúyelo
            al reportar una incidencia.
    ErrorValidacion:
      type: object
      required: [error, code, issues]
      description: |
        Error 422 con el detalle campo a campo: el del cuestionario (v1, v3 o v4) y
        el de la subida de audio (partes del multipart e integridad del archivo).
      properties:
        error:
          type: string
          const: 'Datos inválidos'
        code:
          type: string
          const: VALIDATION_ERROR
        issues:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                description: Ruta del campo, con puntos (`ubicacion.esValida`).
              message:
                type: string
        requestId:
          type: string